+++ src/main/java/com/illtamer/perpetua/sdk/event/request/GroupRequestEvent.java
@@ -73,7 +73,7 @@ public class GroupRequestEvent extends RequestEvent {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Long sendGroupMessage(String message) {
+    public Integer sendGroupMessage(String message) {
         return OpenAPIHandling.sendGroupMessage(message, groupId);
     }
 
@@ -81,7 +81,7 @@ public class GroupRequestEvent extends RequestEvent {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Long sendGroupMessage(Message message) {
+    public Integer sendGroupMessage(Message message) {
         return OpenAPIHandling.sendGroupMessage(message, groupId);
     }
